A few minor changes.
git-svn-id: svn://svn.code.sf.net/p/playerpro/code/trunk@949 f4605116-7794-4d0b-9e48-65ca9bfeedfd

+++ MADDriver.source/MADDriver.Source/MADDriver.h
@@ -69,7 +69,7 @@ struct __MADDriverRec {
 	
 #ifdef _MAC_H
 	AudioUnit				CAAudioUnit;
-	UInt32					CABufOff;
+	size_t					CABufOff;
 	Ptr						CABuffer;
 #endif
